% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumLexc_mul.R
\name{cumLexc_mul}
\alias{cumLexc_mul}
\title{cumLexc_mul function}
\usage{
cumLexc_mul(z_tau, z_alpha, x, theta)
}
\arguments{
\item{z_tau}{covariates depending on tau}

\item{z_alpha}{covariates depending on alpha}

\item{x}{time value}

\item{theta}{of the coefficient of tneh parameters}
}
\value{
An object of class numeric containing the cumulative excess
hazard with the same length as the time.
}
\description{
returns the cumulative excess hazard for an TNEH model
in case of parametrization of log the of the time to null excess hazard as
function to fit the data
}
\keyword{cumLexc_mul}
