% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexhazdtheta_adtneh2.R
\name{dexhazdtheta_adtneh2}
\alias{dexhazdtheta_adtneh2}
\title{dexhazdtheta_adtneh2 function}
\usage{
dexhazdtheta_adtneh2(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  cumLexctopred = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the estimates are predicted}

\item{object}{ouput from model implemented in curesurv}

\item{cumLexctopred}{a pre-prediction parameter, calculated if NULL}
}
\description{
Partial derivatives of excess hazard
by theta from non-mixture model with distribution "tneh".
}
\keyword{internal}
