% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexhazdtheta_wei.R
\name{dexhazdtheta_wei}
\alias{dexhazdtheta_wei}
\title{dexhazdtheta_wei function}
\usage{
dexhazdtheta_wei(
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x = x,
  theta,
  cumLexctopred
)
}
\arguments{
\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{theta}{estimated parameters from a mixture model using curesurv
and uncured survival following a Weibull distribution}

\item{cumLexctopred}{pre prediction obtained with cumLexc_alphaweibull_topred}
}
\description{
Produce partial derivatives of excess hazard
}
\keyword{internal}
