% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digamma.R
\name{digamma}
\alias{digamma}
\title{digamma function}
\usage{
digamma(z)
}
\arguments{
\item{z}{positive argument}
}
\value{
An object of type \code{numeric}.
This object is a vector of the same length as x.
The output element is :

\item{psi }{\eqn{frac{d ln(\tau(x))}{dx} = \frac{\tau'(x)}{\tau(x)}}.
This equal to the output of base::digamma(x)}
}
\description{
digamma function
}
\keyword{internal}
