% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogCumHazdtheta.R
\name{dlogCumHazdtheta_wei}
\alias{dlogCumHazdtheta_wei}
\title{dlogCumHazdtheta_wei function}
\usage{
dlogCumHazdtheta_wei(
  z_ucured = z_ucured,
  z_pcured = z_pcured,
  x = x,
  theta = theta,
  cumLexctopred
)
}
\arguments{
\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{x}{time at which the estimates are predicted}

\item{theta}{estimated parameters from the ouput of a mixture cure model implemented in curesurv}

\item{cumLexctopred}{a pre-prediction parameter obtained with cumLexc_alphaweibull_topred}
}
\description{
function of partial derivates of log-log of net survival depending on
theta (estimated parameters) from a mixture cure model with uncured survival
following a Weibull distribution
}
\keyword{internal}
