% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloglogpdtheta_wei.R
\name{dloglogpdtheta_wei}
\alias{dloglogpdtheta_wei}
\title{dloglogpdtheta_wei function}
\usage{
dloglogpdtheta_wei(
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x = x,
  theta,
  cumLexctopred
)
}
\arguments{
\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{theta}{estimated parameters of the cumulative excess hazard from a mixture
model using curesurv and uncured survival following a Weibull distribution}

\item{cumLexctopred}{a pre-prediction parameter obtained with cumLexc_alphaweibull_topred}
}
\description{
Produce partial derivatives of log(log(p(t)))) the logarithm of the
probability to be cured
}
\keyword{internal}
