% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogTTCdtheta_wei.R
\name{dlogTTCdtheta_wei}
\alias{dlogTTCdtheta_wei}
\title{dlogTTCdtheta_wei function}
\usage{
dlogTTCdtheta_wei(
  z_ucured = z_ucured,
  z_pcured = z_pcured,
  object = object,
  epsilon = epsilon,
  TTC
)
}
\arguments{
\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{object}{ouput from a model implemented in curesurv}

\item{epsilon}{value fixed by user to estimate the TTC \eqn{\text{Pi}(t)\geq (1-\epsilon)}.
By default \eqn{\epsilon = 0.05}.}

\item{TTC}{time to cure previsouly estimated by TTC_wei}
}
\description{
function of partial derivates of log of time-to-cure \code{log(TTC)}
by theta (estimated parameters) from a mixture cure model with uncured survival
following a Weibull distribution
}
\keyword{internal}
