% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdtheta_adtneh2.R
\name{dpdtheta_adtneh2}
\alias{dpdtheta_adtneh2}
\title{dpdtheta_adtneh2 function}
\usage{
dpdtheta_adtneh2(
  z_alpha = z_alpha,
  z_tau = z_tau,
  x = x,
  object,
  cumLexctopred = NULL,
  Dpi = NULL,
  Dsn = NULL
)
}
\arguments{
\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{x}{time at which the estimates are predicted}

\item{object}{ouput from model implemented in curesurv}

\item{cumLexctopred}{a pre-prediction obtained with cumLexc_ad2_topred, if NULL will be calculated}

\item{Dpi}{partial derivatives of pi the cure proportion by theta calculated by dpidtheta_adtneh2, if NULL will be calculated}

\item{Dsn}{partial derivatives of net survival by theta calculated by dsndtheta_adtneh2, if NULL will be calculated}
}
\description{
Partial derivatives of probability to be cure by theta from
non-mixture model with distribution "tneh".
}
\keyword{internal}
