% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdtheta_wei.R
\name{dpdtheta_wei}
\alias{dpdtheta_wei}
\title{dpdtheta_wei function}
\usage{
dpdtheta_wei(
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x = x,
  theta,
  cumLexctopred
)
}
\arguments{
\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{theta}{estimated parameters from a mixture model using curesurv
and uncured survival following a Weibull distribution}

\item{cumLexctopred}{description}
}
\description{
Produce partial derivatives of p(t) the probability to be cured
}
\keyword{internal}
