% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpidtheta_multneh.R
\name{dpidtheta_multneh}
\alias{dpidtheta_multneh}
\title{dpidtheta_multneh function}
\usage{
dpidtheta_multneh(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  cumLexctopred = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the estimates are predicted}

\item{object}{ouput from model implemented in curesurv}

\item{cumLexctopred}{pre prediction obtained from cumLexc_mul_topred, calculated if NULL}
}
\description{
@description Partial derivatives of cure fraction (or net survival at tau)
by theta from non-mixture model with distribution "tneh" when link_tau="loglinear".
}
\keyword{internal}
