% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dptdtheta_multneh.R
\name{dptdtheta_multneh}
\alias{dptdtheta_multneh}
\title{dptdtheta_multneh function}
\usage{
dptdtheta_multneh(
  z_alpha = z_alpha,
  z_tau = z_tau,
  x = x,
  object,
  cumLexctopred = NULL,
  Dpi = NULL,
  Dsn = NULL
)
}
\arguments{
\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{x}{time at which the estimates are predicted}

\item{object}{ouput from model implemented in curesurv}

\item{cumLexctopred}{pre prediction obtained from cumLexc_mul_topred, if NULL will be calculated}

\item{Dpi}{partial derivative of pi according to theta, if NULL will be calculated}

\item{Dsn}{partial derivative of net survival according to theta , if NULL will be calculated}
}
\description{
Partial derivatives of probability to be cure by theta from
non-mixture model with distribution "tneh".
}
\keyword{internal}
