% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTTCdtheta_wei.R
\name{dTTCdtheta_wei}
\alias{dTTCdtheta_wei}
\title{dTTCdtheta_wei function}
\usage{
dTTCdtheta_wei(
  z_ucured = z_ucured,
  z_pcured = z_pcured,
  theta = theta,
  epsilon = epsilon,
  TTC
)
}
\arguments{
\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{theta}{estimated parameters from the ouput of a mixture cure model implemented in curesurv}

\item{epsilon}{value fixed by user to estimate the TTC \eqn{\text{Pi}(t)\geq (1-\epsilon)}.
By default \eqn{\epsilon = 0.05}.}

\item{TTC}{time-to-cure previously estimated using TTC_wei}
}
\description{
function of partial derivates of time-to-cure (TTC) by theta
(estimated parameters) from a mixture cure model with uncured survival
following a Weibull distribution
}
\keyword{internal}
