% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCumHaz_ic_wei.R
\name{logCumHaz_ic_wei}
\alias{logCumHaz_ic_wei}
\title{logCumHaz_ic_wei function}
\usage{
logCumHaz_ic_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x,
  level,
  cumLexctopred
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{a pre prediction obtained with cumLexc_alphaweibull}
}
\description{
calculates the confidence intervals of the log of cumulative excess hazard log(-log(Sn(t)))
using variances of log(-log(Sn(t))). In this formula, the variance of log of
the cumulative excess hazard is obtained using delta method at the scale of log-log of net survival, and with
the expression Var(log(log(Sn(t)))) = (dlog(-log(Sn(t)))/dtheta)Var(theta)(dlog(-log(Sn(t)))/dtheta)^T;
the Var(theta) is the variance-covariance matrix of theta (estimated parameters).
}
\keyword{internal}
