% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancreas_data.R
\docType{data}
\name{pancreas_data}
\alias{pancreas_data}
\title{Simulated pancreas data with vital status information}
\format{
This dataset contains the following variables:
\describe{
\item{age}{Age at diagnosis}
\item{age_cr}{centered and scaled age at diagnosis}
\item{age_classe}{"<45", "45_59" and  ">=60" age groups }
\item{time_obs}{Follow-up time (years)}
\item{event}{Vital status}
\item{cumehazard}{individual cumulative expected hazard}
\item{ehazard}{individual instantaneous expected hazard}
}
}
\usage{
data(pancreas_data)
}
\description{
Simulated data
}
\examples{
data(pancreas_data)
summary(pancreas_data)
}
\keyword{datasets}
