% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcured.R
\name{pcured}
\alias{pcured}
\title{pcured function}
\usage{
pcured(
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object
)
}
\arguments{
\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the predictions are provided}

\item{object}{ouput from a model implemented in curesurv}
}
\value{
An object of class \code{c("cure_fraction", "data.frame")}.
This object is a list containing the following components:

\item{time}{time in the input new data}

\item{netsurv}{predicted net survival at the time provided in the new data}

\item{pi}{pi or net survival at time tau}
}
\description{
calculates cure fraction of mixture and non mixture cure models.
}
\keyword{internal}
\keyword{pcured}
