% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ic_adtneh2_log.R
\name{pi_ic_adtneh2_log}
\alias{pi_ic_adtneh2_log}
\title{pi_ic_adtneh2_log function}
\usage{
pi_ic_adtneh2_log(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  level = level,
  cumLexc_topred = NULL,
  Dpi = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the predictions are provided}

\item{object}{ouput from a model implemented in curesurv}

\item{level}{\code{(1-alpha/2)}-order quantile of a normal distribution}

\item{cumLexc_topred}{a pre prediction, if NULL it is calculated}

\item{Dpi}{partial derivative of pi by theta obtained by dpidtheta_adtneh2 function, if NULL it is calculated}
}
\description{
produces confidence interval of cure fraction pi using a
time-to-null excess hazard model with linear effect on parameter tau
The confidence intervals based on log method.
}
\keyword{internal}
