% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ic_loglog_wei.R
\name{pi_ic_loglog_wei}
\alias{pi_ic_loglog_wei}
\title{pi_ic_loglog_wei function}
\usage{
pi_ic_loglog_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x,
  level,
  cumLexctopred,
  Dpi
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{a pre prediction obtained with cumLexc_alphaweibull_topred}

\item{Dpi}{Partial derivative of Pi calculated by dpidtheta_wei function}
}
\description{
calculates the confidence intervals of the cure proportion pi
using variances of log(-log(pi)) by delta method
}
\keyword{internal}
