% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ic_multneh_loglog.R
\name{pi_ic_multneh_loglog}
\alias{pi_ic_multneh_loglog}
\title{pi_ic_multneh_loglog function}
\usage{
pi_ic_multneh_loglog(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  level = level,
  cumLexctopred = NULL,
  Dpi = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the predictions are provided}

\item{object}{ouput from a model implemented in curesurv}

\item{level}{\code{(1-alpha/2)}-order quantile of a normal distribution}

\item{cumLexctopred}{pre prediction obtained from cumLexc_mul_topred, calculated if NULL}

\item{Dpi}{partial derivative of pi according to theta, if NULL calculated}
}
\description{
produces confidence interval of cure fraction pi using a
time-to-null excess hazard model with log linear effect on parameter tau
The confidence intervals based on log-log method.
}
\keyword{internal}
