% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cure_ic_adtneh2_log.R
\name{pt_cure_ic_adtneh2_log}
\alias{pt_cure_ic_adtneh2_log}
\title{pt_cure_ic_adtneh2_log function}
\usage{
pt_cure_ic_adtneh2_log(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  level = level,
  cumLexctopred = NULL,
  Dpt_cure = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the predictions are provided}

\item{object}{ouput from a model implemented in curesurv}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{a pre-prediction parameter obtained with cumLexc_ad2_topred, if NULL it will be calculated here}

\item{Dpt_cure}{Partial derivatives of the cure propability according to theta, if NULL it will be calculated here}
}
\description{
calculates the probability of the probability Pi(t) of being
cured at a given time t after diagnosis knowing that he/she was alive up to
time t. It also provides the related confidence intervals using log method.
}
\keyword{internal}
