% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cure_ic_log_wei.R
\name{pt_cure_ic_log_wei}
\alias{pt_cure_ic_log_wei}
\title{pt_cure_ic_log_wei function}
\usage{
pt_cure_ic_log_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x,
  level,
  cumLexctopred,
  Dpt_cure
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{a pre-prediction parameter obtained with cumLexc_alphaweibull}

\item{Dpt_cure}{partial derivatives of pt_cure by theta, obtained with dpdtheta_wei function}
}
\description{
confidence intervals of the probability to be cured at time t by
Delta Method on logarithm of P(t)
}
\keyword{internal}
