% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cure_ic_multneh_log.R
\name{pt_cure_ic_multneh_log}
\alias{pt_cure_ic_multneh_log}
\title{pt_cure_ic_multneh_log function}
\usage{
pt_cure_ic_multneh_log(
  z_tau = z_tau,
  z_alpha = z_alpha,
  x = x,
  object,
  level = level,
  cumLexctopred = NULL,
  Dpt_cure = NULL
)
}
\arguments{
\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{x}{time at which the predictions are provided}

\item{object}{ouput from a model implemented in curesurv}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{pre prediction obtained from cumLexc_mul_topred, if NULL will be calculated}

\item{Dpt_cure}{partial derivative of pt_cure according to theta, if NULL will be calculated}
}
\description{
calculates the probability of the probability Pi(t) of being
cured at a given time t after diagnosis knowing that he/she was alive up to
time t. It also provides the related confidence intervals using log method.
}
\keyword{internal}
