% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cure_ic_noDM.R
\name{pt_cure_ic_noDM}
\alias{pt_cure_ic_noDM}
\title{pt_cure_ic_noDM function}
\usage{
pt_cure_ic_noDM(x, z_pcured, t, level = 0.975, theta)
}
\arguments{
\item{x}{fit ouput from curesurv}

\item{z_pcured}{covariates}

\item{t}{time}

\item{level}{level of confidence}

\item{theta}{estimated parameters}
}
\description{
Variance of p(t) without delta method
}
\keyword{internal}
