% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cure_wei.R
\name{pt_cure_wei}
\alias{pt_cure_wei}
\title{pt_cure_wei function}
\usage{
pt_cure_wei(z_pcured = z_pcured, z_ucured = z_ucured, x, theta)
}
\arguments{
\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{theta}{estimated parameters of the cumulative excess hazard from a mixture
model using curesurv and uncured survival following a Weibull distribution}
}
\description{
calculates the probability Pi(t) of being
cured at a given time t after diagnosis knowing that he/she was alive up to
time t. The predictions are based on a mixture cure model with weibull
distribution for the survival of uncured patients.
}
\keyword{internal}
