% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_ic_log_wei.R
\name{sn_ic_log_wei}
\alias{sn_ic_log_wei}
\title{sn_ic_log_wei function}
\usage{
sn_ic_log_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x,
  level,
  cumLexctopred,
  Dsn
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{a pre prediction obtained with cumLexc_alphaweibull_topred}

\item{Dsn}{Partial derivative of Sn calculated by dsndtheta_wei function}
}
\description{
calculates the confidence intervals of the net survival (Sn(t))
using variances of log(Sn(t)). In this formula, the variance of net survival
is obtained using delta method at the scale of log of net survival, and with
the expression Var(log(Sn(t))) = (dlog(Sn)/dtheta)Var(theta)(dlog(Sn)/dtheta)^T;
the Var(theta) is the variance-covariance matrix of theta (estimated parameters).
}
\keyword{internal}
