% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_ic_wei.R
\name{sn_ic_wei}
\alias{sn_ic_wei}
\title{sn_ic_wei function}
\usage{
sn_ic_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x,
  level,
  cumLexctopred,
  Dsn
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{level}{(1-alpha/2)-order quantile of a normal distribution}

\item{cumLexctopred}{pre prediction obtained with cumLexc_alphaweibull_topred}

\item{Dsn}{Partial derivative of Sn calculated by dsndtheta_wei function}
}
\description{
calculates the confidence intervals of the net survival (Sn(t))
using variances of Sn(t). In this formula, the variance of net survival
is obtained using the expression Var(Sn(t)) = (dSn/dtheta)Var(theta)(dSn/dtheta)^T
where Var(theta) is the variance-covariance matrix of theta.
}
\keyword{internal}
