% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTC_ic_Jakobsen_wei.R
\name{TTC_ic_Jakobsen_wei}
\alias{TTC_ic_Jakobsen_wei}
\title{TTC_ic_Jakobsen_wei function}
\usage{
TTC_ic_Jakobsen_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  epsilon = 0.05,
  level = 0.975
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{epsilon}{value fixed by user to estimate the TTC \eqn{\text{Pi}(t)\geq (1-\epsilon)}.
By default \eqn{\epsilon = 0.05}.}

\item{level}{\code{1-alpha/2}-order quantile of a normal distribution}
}
\description{
calculates the confidence interval of the time TTC using the
Jakobsen's approach.
Note that this function is for mixture cure model with Weibull distribution
considered for uncured patients.
}
\keyword{internal}
