% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTC_theoric_adtneh.R
\name{TTC_theoric_adtneh}
\alias{TTC_theoric_adtneh}
\title{TTC_theoric_adtneh function}
\usage{
TTC_theoric_adtneh(newdata, object)
}
\arguments{
\item{newdata}{newdata}

\item{object}{An object of class \code{curesurv}.}
}
\description{
Product the value of TTC with TNEH model
}
\keyword{internal}
