% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_logCumHaz_wei.R
\name{var_logCumHaz_wei}
\alias{var_logCumHaz_wei}
\title{var_logCumHaz_wei function}
\usage{
var_logCumHaz_wei(
  object,
  z_ucured = z_ucured,
  z_pcured = z_pcured,
  x = x,
  cumLexctopred
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{x}{time at which the estimates are predicted}

\item{cumLexctopred}{pre prediction obtained with cumLexc_alphaweibull_topred}
}
\description{
Calculates the variance of log of cumulative excess hazard with
delta method on the log-log of net survival scale.
}
\keyword{internal}
