% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_pt_cure_wei.R
\name{var_pt_cure_wei}
\alias{var_pt_cure_wei}
\title{var_pt_cure_wei function}
\usage{
var_pt_cure_wei(
  object,
  z_pcured = z_pcured,
  z_ucured = z_ucured,
  x = x,
  cumLexctopred
)
}
\arguments{
\item{object}{ouput from mixture cure model implemented in curesurv}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{x}{time at which the estimates are predicted}

\item{cumLexctopred}{pre prediction obtained with cumLexc_alphaweibull_topred}
}
\description{
Variance of p(t) with delta method. Var(p(t)) = (dp/dtheta)Var(theta)(dp/dtheta)^T
where Var(theta) is the variance-covariance matrix of theta.
}
\keyword{internal}
