% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_TTC_multneh.R
\name{var_TTC_multneh}
\alias{var_TTC_multneh}
\title{var_TTC_multneh function}
\usage{
var_TTC_multneh(
  z_alpha,
  z_tau,
  xmax,
  object,
  epsilon = epsilon,
  TTC = NULL,
  cumLexctopred = NULL,
  DpTTC = NULL
)
}
\arguments{
\item{z_alpha}{Covariates matrix acting on parameter alpha of the density of
time-to-null excess hazard model}

\item{z_tau}{Covariates matrix acting on time-to-null parameter.}

\item{xmax}{time max at which Pi(t) is calculated.}

\item{object}{ouput from a model implemented in curesurv}

\item{epsilon}{value fixed by user to estimate the TTC \eqn{\text{Pi}(t)\geq (1-\epsilon)}.
By default \eqn{\epsilon = 0.05}.}

\item{TTC}{The time to cure, if NULL it is recalculated}

\item{cumLexctopred}{pre prediction, calculated if NULL}

\item{DpTTC}{partial derivatives, recalculated if not given}
}
\description{
calculates the variance of TTC in a non-mixture model with
distribution "TNEH", link_tau="loglinear"
}
\references{
Boussari O, Bordes L, Romain G, Colonna M, Bossard N, Remontet L,
Jooste V. Modeling excess hazard with time-to-cure as a parameter.
Biometrics. 2021 Dec;77(4):1289-1302. doi: 10.1111/biom.13361.
Epub 2020 Sep 12. PMID: 32869288.
(\href{https://pubmed.ncbi.nlm.nih.gov/32869288/}{pubmed})

Boussari O, Romain G, Remontet L, Bossard N, Mounier M, Bouvier AM,
Binquet C, Colonna M, Jooste V. A new approach to estimate time-to-cure from
cancer registries data. Cancer Epidemiol. 2018 Apr;53:72-80.
doi: 10.1016/j.canep.2018.01.013. Epub 2018 Feb 4. PMID: 29414635.
(\href{https://pubmed.ncbi.nlm.nih.gov/29414635/}{pubmed})
}
\author{
Juste Goungounga, Judith Breaud, Eugenie Blandin, Olayidé Boussari, Valérie Jooste
}
\keyword{internal}
