% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlogTTC_Jakobsen_wei.R
\name{varlogTTC_Jakobsen_wei}
\alias{varlogTTC_Jakobsen_wei}
\title{varlogTTC_Jakobsen_wei function}
\usage{
varlogTTC_Jakobsen_wei(
  object,
  z_ucured = z_ucured,
  z_pcured = z_pcured,
  epsilon = epsilon,
  TTC,
  cumLexc_topred_TTC
)
}
\arguments{
\item{object}{ouput from a model implemented in curesurv}

\item{z_ucured}{covariates matrix acting on survival function of uncured}

\item{z_pcured}{covariates matrix acting on cure proportion}

\item{epsilon}{value fixed by user to estimate the TTC \eqn{\text{Pi}(t)\geq (1-\epsilon)}.
By default \eqn{\epsilon = 0.05}.}

\item{TTC}{time to cure estimatec by TTC_wei}

\item{cumLexc_topred_TTC}{a pre prediction given by cumLexc_alphaweibull_topred}
}
\description{
Calculates the variance of log(TTC) from a mixture cure model with
uncured survival as Weibull distribution using the Jakobsen approach.
}
\keyword{internal}
