% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_alpha.R
\name{z_alpha}
\alias{z_alpha}
\title{z_alpha function identifying variables acting on alpha parameter}
\usage{
z_alpha(x)
}
\arguments{
\item{x}{a simple formula.}
}
\value{
the variable x
}
\description{
variables adjusted on alpha parameter in
non-mixture cure model with "tneh" specified for the distribution.
}
\references{
Boussari O, Bordes L, Romain G, Colonna M, Bossard N,
Remontet L, Jooste V. Modeling excess hazard with time-to-cure
as a parameter. Biometrics. 2020 Aug 31.
doi: 10.1111/biom.13361. Epub ahead of print.
PMID: 32869288.
(\href{https://pubmed.ncbi.nlm.nih.gov/32869288/}{pubmed})
}
\author{
Juste Goungounga, Judith Breaud, Olayide Boussari, Gaelle Romain, Valerie Jooste
}
\keyword{z_alpha}
