% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_riskpop.alpha.R
\name{z_riskpop.alpha}
\alias{z_riskpop.alpha}
\title{z_riskpop.alpha function}
\usage{
z_riskpop.alpha(x)
}
\arguments{
\item{x}{a simple formula.}
}
\value{
the variable x
}
\description{
indicator variable.
}
\references{
Boussari O, Bordes L, Romain G, Colonna M, Bossard N,
Remontet L, Jooste V. Modeling excess hazard with time-to-cure
as a parameter. Biometrics. 2020 Aug 31.
doi: 10.1111/biom.13361. Epub ahead of print.
PMID: 32869288.
(\href{https://pubmed.ncbi.nlm.nih.gov/32869288/}{pubmed})
}
\author{
Juste Goungounga, Judith Breaud, Olayide Boussari, Gaelle Romain, Valerie Jooste
}
\keyword{internal}
\keyword{z_riskpop.alpha}
