% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_tau.R
\name{z_tau}
\alias{z_tau}
\title{z_tau function identifying variables acting on tau parameter}
\usage{
z_tau(x)
}
\arguments{
\item{x}{the name of the column in the dataset representing the variable that will act on tau parameter of the "tneh" model}
}
\value{
the variable x
}
\description{
variables adjusted on tau parameter in
non-mixture cure model with "tneh" specified for the distribution.
}
\references{
Boussari O, Bordes L, Romain G, Colonna M, Bossard N,
Remontet L, Jooste V. Modeling excess hazard with time-to-cure
as a parameter. Biometrics. 2020 Aug 31.
doi: 10.1111/biom.13361. Epub ahead of print.
PMID: 32869288.
(\href{https://pubmed.ncbi.nlm.nih.gov/32869288/}{pubmed})
}
\author{
Juste Goungounga, Judith Breaud, Eugenie Blandin, Olayide Boussari, Valerie Jooste
}
\keyword{z_tau}
